+++ android/src/main/java/com/revenuecat/purchases/react/RNPurchasesModule.java
@@ -49,8 +49,9 @@ public class RNPurchasesModule extends ReactContextBaseJavaModule implements Upd
     private static final String CUSTOMER_INFO_UPDATED = "Purchases-CustomerInfoUpdated";
     private static final String LOG_HANDLER_EVENT = "Purchases-LogHandlerEvent";
     private static final String TRACKED_EVENT = "Purchases-TrackedEvent";
+    private static final String DEBUG_EVENT = "Purchases-DebugEvent";
     public static final String PLATFORM_NAME = "react-native";
-    public static final String PLUGIN_VERSION = "9.9.0";
+    public static final String PLUGIN_VERSION = "9.10.1";
 
     private final ReactApplicationContext reactContext;
 
@@ -298,6 +299,15 @@ public class RNPurchasesModule extends ReactContextBaseJavaModule implements Upd
         CommonKt.setTrackedEventListener(eventDetails -> {
             reactContext.getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
                 .emit(RNPurchasesModule.TRACKED_EVENT, convertMapToWriteableMap(eventDetails));
+            return null;
+        });
+    }
+
+    @ReactMethod
+    public void setDebugEventListener() {
+        CommonKt.setDebugEventListener(eventDetails -> {
+            reactContext.getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
+                .emit(RNPurchasesModule.DEBUG_EVENT, convertMapToWriteableMap(eventDetails));
             return null;
         });
     }