+++ react-native-purchases-ui/src/index.tsx
@@ -80,6 +80,7 @@ const InternalPaywall: React.FC<FullScreenPaywallViewProps> = ({
   onRestoreCompleted,
   onRestoreError,
   onDismiss,
+  onPurchasePackageInitiated,
 }) => {
   if (usingPreviewAPIMode) {
     return (
@@ -113,6 +114,17 @@ const InternalPaywall: React.FC<FullScreenPaywallViewProps> = ({
         onRestoreCompleted={(event: any) => onRestoreCompleted && onRestoreCompleted(event.nativeEvent)}
         onRestoreError={(event: any) => onRestoreError && onRestoreError(event.nativeEvent)}
         onDismiss={() => onDismiss && onDismiss()}
+        onPurchasePackageInitiated={(event: any) => {
+          const { packageBeingPurchased, requestId } = event.nativeEvent;
+          if (onPurchasePackageInitiated) {
+            const resume = (shouldProceed: boolean) => {
+              RNPaywalls!.resumePurchasePackageInitiated(requestId, shouldProceed);
+            };
+            onPurchasePackageInitiated({ packageBeingPurchased, resume });
+          } else {
+            RNPaywalls!.resumePurchasePackageInitiated(requestId, true);
+          }
+        }}
       />
     );
   }
@@ -287,6 +299,10 @@ type FullScreenPaywallViewProps = {
   onRestoreCompleted?: ({customerInfo}: { customerInfo: CustomerInfo }) => void;
   onRestoreError?: ({error}: { error: PurchasesError }) => void;
   onDismiss?: () => void;
+  onPurchasePackageInitiated?: ({
+    packageBeingPurchased, 
+    resume
+  }: { packageBeingPurchased: PurchasesPackage, resume: (shouldResume: boolean) => void}) => void;
 };
 
 type FooterPaywallViewProps = {
@@ -509,6 +525,7 @@ export default class RevenueCatUI {
                                                                    onRestoreCompleted,
                                                                    onRestoreError,
                                                                    onDismiss,
+                                                                   onPurchasePackageInitiated,
                                                                  }) => {
     return (
       <InternalPaywall
@@ -522,6 +539,7 @@ export default class RevenueCatUI {
         onRestoreCompleted={onRestoreCompleted}
         onRestoreError={onRestoreError}
         onDismiss={onDismiss}
+        onPurchasePackageInitiated={onPurchasePackageInitiated}
         style={[{flex: 1}, style]}
       />
     );