+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## 9.10.1
+## RevenueCat SDK
+### 📦 Dependency Updates
+* [AUTOMATIC BUMP] Updates purchases-hybrid-common to 17.40.0 (#1608) via RevenueCat Git Bot (@RCGitBot)
+  * [Android 9.22.0](https://github.com/RevenueCat/purchases-android/releases/tag/9.22.0)
+
+### 🔄 Other Changes
+* Add debugEventListener internal API (#1609) via Toni Rico (@tonidero)
+
+## 9.10.0
+## RevenueCat SDK
+### ✨ New Features
+* [SDK-4254] Add onPurchasePackageInitiated callback to Paywall component (#1598) via Toni Rico (@tonidero)
+### 🐞 Bugfixes
+* Fix: footer paywall purchase hang when onPurchasePackageInitiated is not set (iOS) (#1605) via Toni Rico (@tonidero)
+### 📦 Dependency Updates
+* [AUTOMATIC BUMP] Updates purchases-hybrid-common to 17.38.0 (#1600) via RevenueCat Git Bot (@RCGitBot)
+  * [iOS 5.59.0](https://github.com/RevenueCat/purchases-ios/releases/tag/5.59.0)
+  * [iOS 5.58.1](https://github.com/RevenueCat/purchases-ios/releases/tag/5.58.1)
+
+## RevenueCatUI SDK
+### ✨ New Features
+* Add custom variables support for paywalls (#1581) via Facundo Menzella (@facumenzella)
+
 ## 9.9.0
 ## RevenueCat SDK
 ### ✨ New Features